+++ Assets/Resources/Prefabs/UI/LoadMapScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5451bfc8a3510da4a96c753c86ee91f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
