+++ Assets/Resources/Scripts/Map/SaveHandler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 492857f3ee77fea478d464b9f82302ba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
