+++ Assets/Resources/Scripts/SaveInfo/Save.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c1c5e17a3dc75eb41b2be7d18dd5a137
+guid: b66e16a765bcdc94a9c1fe350052399c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
