+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f5370eabcdb43f44b9052800ba43479
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
