+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedBoard.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using UnityEngine;
+public class NewBehaviourScript1 : MonoBehaviour
+{
+
+    // Use this for initialization
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+}
