+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedBoard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ccdd9abbe74d3c43b4a095461779697
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
