+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedColor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e27c6c1caf4d1764cb578a72af020743
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
