+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedCulture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a66eec0f364957342aed58fa259d0762
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
