+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedCultureMemory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 83fee9e1b2840f045a7d733807e3da42
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
