+++ Assets/Resources/Scripts/SaveInfo/SerializedClasses/SerializedTile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bae07ff2d40471b45b00a1f401093a78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
