+++ Assets/Resources/Scripts/UI/LoadMapScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e3a3ea9ec7a5ec84abcf996a6141d610
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
